README with note about temporary suspension
Work on this project is suspended until @brixen finishes this batch of updates to `Rubinius::Melbourne` and `Rubinius::AST`.  This should be within the next few months.  Once the AST changes are complete and the specs are settled, this project will finish adapting Parser to generate the same AST objects and work on integration via `Rubinius::ToolSets`.